requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
attrs==24.3.0
bertopic==0.16.4
blinker==1.9.0
blis==1.2.0
bump-pydantic==0.8.0
catalogue==2.0.10
category_encoders==2.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
cloudpickle==3.1.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
Cython==3.0.11
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2024.12.1
datallm==0.2.6
datasets==3.2.0
debugpy==1.8.11
decorator==5.1.1
deprecation==2.1.0
dill==0.3.8
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl#sha256=293e9547a655b25499198ab15a525b05b9407a75f10255e405e8c3854329ab63
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
fuzzywuzzy==0.18.0
gast==0.6.0
git-filter-repo==2.47.0
google-pasta==0.2.0
grpcio==1.69.0
h11==0.14.0
h5py==3.12.1
hdbscan==0.8.40
htmlmin==0.1.12
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
ImageHash==4.3.1
imbalanced-learn==0.13.0
importlib_metadata==8.5.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kaleido==0.2.1
keras==3.8.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
Levenshtein==0.26.1
libclang==18.1.1
libcst==1.6.0
lightgbm==4.5.0
llvmlite==0.43.0
locket==1.0.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
missingno==0.5.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
multimethod==2.0
multiprocess==0.70.16
murmurhash==1.0.12
mysql-connector-python==9.2.0
namex==0.0.8
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.14.0
orjson==3.10.14
packaging==24.2
pandas==2.1.4
parso==0.8.4
partd==1.4.2
patsy==1.0.1
pexpect==4.9.0
phik==0.12.4
pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
plotly-resampler==0.10.0
pluggy==1.5.0
pmdarima==2.0.4
preshed==3.0.9
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyarrow==18.1.0
pycaret==3.3.2
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pynndescent==0.5.13
pyod==2.0.3
pyparsing==3.2.1
pytest==6.2.5
pytest-timeout==2.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.1
pytz==2024.2
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.11.0
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.2
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
seaborn==0.13.2
sentence-transformers==3.4.1
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
sktime==0.26.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.37
srsly==2.5.1
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.13.1
tangled-up-in-unicode==0.2.0
tbats==1.1.3
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
thinc==8.3.4
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
toolz==1.0.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.0
tsdownsample==0.1.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
ucimlrepo==0.0.7
umap-learn==0.5.7
urllib3==2.3.0
visions==0.7.4
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.0.6
wheel==0.45.1
widgetsnbextension==4.0.13
wrapt==1.17.2
wurlitzer==3.1.1
xgboost==2.1.3
xxhash==3.5.0
yarl==1.18.3
yellowbrick==1.5
zipp==3.21.0
